continue working on 5-day cards
