+++ electron/styles/weather.css
@@ -174,6 +174,24 @@
   display: inline-block;
 }
 
+.temperature-display-feels-like-daily {
+  font-size: 10.5px;
+  display: inline-block;
+}
+
+.morning-evening-night-enclosure {
+  width: 100%;
+}
+
+.cat {
+  color: yellow;
+  font-weight: bold;
+}
+
+.daily-temp {
+  font-size: 9.5px;
+}
+
 .forecast-cards-enclosure {
   padding-left: 1.1em;
   padding-right: 1.1em;
